Move grammar to proper translation unit & write tests
