+++ source/mdb/mdbsql/test/mdbsql_engine.t.cpp
@@ -8,13 +8,13 @@ using namespace melinda;
 
 TEST_CASE("Creating a database succeeds", "[exec]")
 {
-    //    mdbsql::engine engine {mdbsql::t::runtime_directory()};
-    //    REQUIRE(engine.execute("CREATE DATABASE books;"));
+    mdbsql::engine engine {mdbsql::t::runtime_directory()};
+    REQUIRE(engine.execute("CREATE SCHEMA books;"));
 }
 
 TEST_CASE("Creating a already existing dabase fails", "[exec]")
 {
-    //    mdbsql::engine engine {mdbsql::t::runtime_directory()};
-    //    REQUIRE(engine.execute("CREATE DATABASE books;"));
-    //    REQUIRE_FALSE(engine.execute("CREATE DATABASE books;"));
+    mdbsql::engine engine {mdbsql::t::runtime_directory()};
+    REQUIRE(engine.execute("CREATE SCHEMA books;"));
+    REQUIRE_FALSE(engine.execute("CREATE SCHEMA books;"));
 }
